-add new test for method telephony.externalCall.searchCrmEntities
