+++ tests/Integration/Services/Telephony/Service/ExternalCallTest.php
@@ -7,6 +7,7 @@ namespace Bitrix24\SDK\Tests\Integration\Services\Telephony\Service;
 
 use Bitrix24\SDK\Core\Exceptions\BaseException;
 use Bitrix24\SDK\Core\Exceptions\TransportException;
+use Bitrix24\SDK\Services\CRM\Contact\Service\Contact;
 use Bitrix24\SDK\Services\CRM\Lead\Service\Lead;
 use Bitrix24\SDK\Services\Main\Service\Main;
 use Bitrix24\SDK\Services\Telephony\Common\CallType;
@@ -24,6 +25,7 @@ class ExternalCallTest extends TestCase
     protected Lead $leadService;
     protected ExternalCall $externalCallService;
     private Main $mainService;
+    protected Contact $contactService;
 
     /**
      * @throws BaseException
@@ -48,7 +50,6 @@ class ExternalCallTest extends TestCase
                 ]
             ]
         )->getId();
-
         $userId = $this->mainService->getCurrentUserProfile()->getUserProfile()->ID;
         $registerCallResult = $this->externalCallService->registerCall([
             'USER_PHONE_INNER' => '14',
@@ -62,7 +63,7 @@ class ExternalCallTest extends TestCase
             'SHOW' => 1,
             'CALL_LIST_ID' => 1,
             'LINE_NUMBER' => $phoneNumber,
-            'TYPE' => CallType::inboundCall(),
+            'TYPE' => (string)CallType::inboundCall(),
         ])->getExternalCallRegister();
 
         self::assertTrue((bool)$registerCallResult);
@@ -182,7 +183,7 @@ class ExternalCallTest extends TestCase
             'CALL_START_DATE' => $callStartDate,
             'CRM_CREATE' => 0,
             'CRM_SOURCE' => '1',
-            'CRM_ENTITY_TYPE' => 'LEAD',
+            'CRM_ENTITY_TYPE' => CrmEntityType::lead(),
             'CRM_ENTITY_ID' => $leadId,
             'SHOW' => 1,
             'CALL_LIST_ID' => 1,
@@ -245,7 +246,7 @@ class ExternalCallTest extends TestCase
             'CALL_START_DATE' => $callStartDate,
             'CRM_CREATE' => 0,
             'CRM_SOURCE' => '1',
-            'CRM_ENTITY_TYPE' => 'LEAD',
+            'CRM_ENTITY_TYPE' => CrmEntityType::lead(),
             'CRM_ENTITY_ID' => $leadId,
             'SHOW' => 1,
             'CALL_LIST_ID' => 1,
@@ -259,7 +260,7 @@ class ExternalCallTest extends TestCase
             'DURATION' => 10,
             'COST' => 250,
             'COST_CURRENCY' => 'RUB',
-            'STATUS_CODE' => 'VI_STATUS_200',
+            'STATUS_CODE' => StatusSipCodeInterface::STATUS_OK,
             'FAILED_REASON' => '',
             'RECORD_URL' => '',
             'VOTE' => 5,
@@ -268,63 +269,6 @@ class ExternalCallTest extends TestCase
 
         $fileName = sprintf('test%s.mp3', time());
         self::assertGreaterThan(1, $this->externalCallService->attachRecord($registerCallResult->CALL_ID, $fileName, $this->getFileInBase64())->getFileId());
-    }
-
-    /**
-     * @throws TransportException
-     * @throws BaseException
-     * @throws Exception
-     * @covers ExternalCall::searchCrmEntities
-     */
-    public function testSearchCrmEntities():void
-    {
-        $datetime = new DateTime('now');
-        $callStartDate = $datetime->format(DateTimeInterface::ATOM);
-        $phoneNumber = sprintf('+7%s', time());
-        $leadId = $this->leadService->add(
-            [
-                'TITLE' => 'test lead',
-                'PHONE' => [
-                    [
-                        'VALUE' => $phoneNumber,
-                        'VALUE_TYPE' => 'WORK'
-                    ]
-                ]
-            ]
-        )->getId();
-        $userId = $this->mainService->getCurrentUserProfile()->getUserProfile()->ID;
-        $registerCallResult = $this->externalCallService->registerCall([
-            'USER_PHONE_INNER' => '14',
-            'USER_ID' => $userId,
-            'PHONE_NUMBER' => $phoneNumber,
-            'CALL_START_DATE' => $callStartDate,
-            'CRM_CREATE' => 0,
-            'CRM_SOURCE' => '1',
-            'CRM_ENTITY_TYPE' => 'LEAD',
-            'CRM_ENTITY_ID' => $leadId,
-            'SHOW' => 1,
-            'CALL_LIST_ID' => 1,
-            'LINE_NUMBER' => $phoneNumber,
-            'TYPE' => (string)CallType::inboundCall(),
-        ])->getExternalCallRegister();
-
-        $finishCallResult = $this->externalCallService->finish([
-            'CALL_ID' => $registerCallResult->CALL_ID,
-            'USER_ID' => $userId,
-            'DURATION' => 255,
-            'COST' => 250,
-            'COST_CURRENCY' => 'RUB',
-            'STATUS_CODE' => StatusSipCodeInterface::STATUS_OK,
-            'FAILED_REASON' => '',
-            'RECORD_URL' => '',
-            'VOTE' => 5,
-            'ADD_TO_CHAT' => 1
-        ])->getExternalCallFinish();
-
-
-        $infoAboutClientResult = $this->externalCallService->searchCrmEntities($phoneNumber)->getCrmEntitiesClient();
-        self::assertNotEmpty($phoneNumber);
-        self::assertTrue((bool)$infoAboutClientResult);
     }
 
     /**
